Отрефакторит константы.
