+++ src/const/filter.js
@@ -1,0 +1,18 @@
+const FILTER_TYPE = {
+  EVERYTHING: 'everything',
+  FUTURE: 'future',
+  PRESENT: 'present',
+  PAST: 'past',
+};
+
+const NoPointsTextType = {
+  [FILTER_TYPE.EVERYTHING]: 'Click New Event to create your first point',
+  [FILTER_TYPE.FUTURE]: 'There are no future events now',
+  [FILTER_TYPE.PRESENT]: 'There are no present events now',
+  [FILTER_TYPE.PAST]: 'There are no past events now',
+};
+
+export {
+  FILTER_TYPE,
+  NoPointsTextType
+};
