+++ src/const/sort.js
@@ -1,0 +1,24 @@
+const SORT_TYPE = {
+  DAY: {
+    name: 'day',
+    enabled: true,
+  },
+  EVENT: {
+    name: 'event',
+    enabled: false,
+  },
+  TIME: {
+    name: 'time',
+    enabled: true,
+  },
+  PRICE: {
+    name: 'price',
+    enabled: true,
+  },
+  OFFERS: {
+    name: 'offers',
+    enabled: false,
+  },
+};
+
+export {SORT_TYPE};
